+++ R/aet05.R
@@ -3,6 +3,7 @@
 #' @describeIn aet05 Main TLG function
 #'
 #' @inheritParams gen_args
+#' @param dataset (`string`) the name of a table in the `adam_db` object.
 #' @param arm_var (`string`) the arm variable used for arm splitting.
 #' @param ... Further arguments passed to `tern::control_incidence_rate()`.
 #'
@@ -11,29 +12,36 @@
 #'  * Split columns by arm, typically `ACTARM`.
 #'  * Split rows by parameter code.
 #'  * `AVAL` is patient-years at risk.
-#'  * `n_events` is the number of adverse events observed.
+#'  * `N_EVENTS` is the number of adverse events observed.
 #'  * The table allows confidence level to be adjusted, default is 95%.
 #'  * Keep zero count rows by default.
 #'
 #' @note
-#'  * `adam_db` object must contain an `adaette` table with the columns `"PARAMCD"`, `"PARAM"`, `"AVAL"`, and `"CNSR"`.
+#'  * `adam_db` object must contain table named as `dataset` with the columns `"PARAMCD"`, `"PARAM"`,
+#'  `"AVAL"`, and `"CNSR"`.
 #'
 #' @export
 #'
 aet05_main <- function(adam_db,
+                       dataset = "adsaftte",
                        arm_var = "ACTARM",
                        lbl_overall = NULL,
                        ...) {
-  assert_all_tablenames(adam_db, c("adsl", "adaette"))
+  assert_string(dataset)
+  assert_all_tablenames(adam_db, "adsl", dataset)
   assert_string(arm_var)
   assert_string(lbl_overall, null.ok = TRUE)
+  df_lbl <- paste0("adam_db$", dataset)
   assert_valid_variable(adam_db$adsl, c("USUBJID", arm_var), types = list(c("character", "factor")))
-  assert_valid_variable(adam_db$adaette, c("USUBJID", arm_var, "PARAMCD", "PARAM"),
-    types = list(c("character", "factor"))
+  assert_valid_variable(adam_db[[dataset]], c("USUBJID", arm_var, "PARAMCD", "PARAM"),
+    types = list(c("character", "factor")), label = df_lbl
   )
-  assert_valid_variable(adam_db$adaette, "AVAL", types = list("numeric"), lower = 0, na_ok = TRUE)
-  assert_valid_variable(adam_db$adaette, "n_events", types = list("numeric"), integerish = TRUE, lower = 0L)
-  assert_valid_var_pair(adam_db$adsl, adam_db$adaette, arm_var)
+  assert_valid_variable(adam_db[[dataset]], "AVAL", types = list("numeric"), lower = 0, na_ok = TRUE, label = df_lbl)
+  assert_valid_variable(adam_db[[dataset]], "N_EVENTS",
+    types = list("numeric"), integerish = TRUE, lower = 0L,
+    label = df_lbl
+  )
+  assert_valid_var_pair(adam_db$adsl, adam_db[[dataset]], arm_var)
 
   lbl_overall <- render_safe(lbl_overall)
   control <- execute_with_args(control_incidence_rate, ...)
@@ -43,11 +51,11 @@ aet05_main <- function(adam_db,
     lbl_overall = lbl_overall,
     param_label = "PARAM",
     vars = "AVAL",
-    n_events = "n_events",
+    n_events = "N_EVENTS",
     control = control
   )
 
-  tbl <- build_table(lyt, adam_db$adaette, alt_counts_df = adam_db$adsl)
+  tbl <- build_table(lyt, adam_db[[dataset]], alt_counts_df = adam_db$adsl)
 
   tbl
 }
@@ -85,11 +93,11 @@ aet05_lyt <- function(arm_var,
 #'
 #' @export
 #'
-aet05_pre <- function(adam_db, ...) {
-  adam_db$adaette <- adam_db$adaette %>%
-    filter(grepl("AETTE", .data$PARAMCD)) %>%
+aet05_pre <- function(adam_db, dataset = "adsaftte", ...) {
+  adam_db[[dataset]] <- adam_db[[dataset]] %>%
+    filter(grepl("(AE|CQ|SMQ)TTE", .data$PARAMCD)) %>%
     mutate(
-      n_events = as.integer(.data$CNSR == 0)
+      N_EVENTS = as.integer(.data$CNSR == 0)
     )
 
   adam_db
@@ -120,7 +128,7 @@ aet05_post <- function(tlg, prune_0 = FALSE, ...) {
 #' library(dplyr)
 #' library(dunlin)
 #'
-#' proc_data <- log_filter(syn_data, PARAMCD == "AETTE1", "adaette")
+#' proc_data <- log_filter(syn_data, PARAMCD == "AETTE1", "adsaftte")
 #'
 #' run(aet05, proc_data)
 #'
@@ -128,6 +136,5 @@ aet05_post <- function(tlg, prune_0 = FALSE, ...) {
 aet05 <- chevron_t(
   main = aet05_main,
   preprocess = aet05_pre,
-  postprocess = aet05_post,
-  adam_datasets = c("adsl", "adaette")
+  postprocess = aet05_post
 )