+++ tests/testthat/test-aet05_all.R
@@ -1,20 +1,20 @@
 test_that("aet05_all can handle NA values", {
   proc_data <- syn_data
-  proc_data$adaette$AVAL <- NA_real_
-  proc_data$adaette$CNSR <- NA_real_
+  proc_data$adsaftte$AVAL <- NA_real_
+  proc_data$adsaftte$CNSR <- NA_real_
 
   res <- expect_silent(run(aet05_all, proc_data))
   expect_snapshot(cat(export_as_txt(res, lpp = 100)))
 })
 
 test_that("aet05_all can handle some NA values", {
-  new_paramcd <- c(NA, "", as.character(syn_data$adaette$PARAMCD[-c(1, 2)]))
+  new_paramcd <- c(NA, "", as.character(syn_data$adsaftte$PARAMCD[-c(1, 2)]))
 
   proc_data <- syn_data
 
-  proc_data$adaette <- proc_data$adaette %>%
+  proc_data$adsaftte <- proc_data$adsaftte %>%
     mutate(
-      PARAMCD = with_label(.env$new_paramcd, var_labels_for(syn_data$adaette, "PARAMCD"))
+      PARAMCD = with_label(.env$new_paramcd, var_labels_for(syn_data$adsaftte, "PARAMCD"))
     )
 
   res1 <- expect_silent(run(aet05_all, proc_data))